Another fix for issue #269 - setting of piecewise spindle linearisation values not working. Fix for incorrect reporting of SD card size.

+++ vfs.h
@@ -92,8 +92,8 @@ typedef struct {
 } vfs_dirent_t;
 
 typedef struct {
-    size_t size;
-    size_t used;
+    uint64_t size;
+    uint64_t used;
 } vfs_free_t;
 
 typedef struct {
